+++ src/app/page-component/copy-button/copy-button.component.ts
@@ -1,0 +1,36 @@
+import {Component, Input, OnInit} from '@angular/core';
+
+@Component({
+  selector: 'app-copy-button',
+  templateUrl: './copy-button.component.html',
+  styleUrls: ['./copy-button.component.css']
+})
+export class CopyButtonComponent implements OnInit {
+  @Input() name;
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+  copyTextToClipboard() {
+    const txtArea = document.createElement('textarea');
+    txtArea.id = 'txt';
+    txtArea.style.position = 'fixed';
+    txtArea.style.top = '0';
+    txtArea.style.left = '0';
+    txtArea.style.opacity = '0';
+    txtArea.value = this.name;
+    document.body.appendChild(txtArea);
+    txtArea.select();
+    try {
+      const successful = document.execCommand('copy');
+      if (successful) {
+        return true;
+      }
+    } catch (err) {
+    } finally {
+      document.body.removeChild(txtArea);
+    }
+    return false;
+  }
+}