+++ Changelog.md
@@ -3,6 +3,7 @@
 ### Added
     * Loading while case is delete
     * Add "note" attr for image: annotation(:string) of image
+    * Add copy button to all result_sets
 ### Fixed
     * Fixed race condition while plans load
 ### Changed